use -1 as proposer for missed blocks

+++ handlers/validator.go
@@ -1398,12 +1398,12 @@ func ValidatorSlashings(w http.ResponseWriter, r *http.Request) {
 	for _, b := range attesterSlashings {
 
 		inter := intersect.Simple(b.Attestestation1Indices, b.Attestestation2Indices)
-		slashedValidators := []uint64{}
+		slashedValidators := []int64{}
 		if len(inter) == 0 {
 			logger.Warning("No intersection found for attestation violation")
 		}
 		for _, v := range inter {
-			slashedValidators = append(slashedValidators, uint64(v.(int64)))
+			slashedValidators = append(slashedValidators, v.(int64))
 		}
 
 		tableData = append(tableData, []interface{}{
